requirements.txt
requests~=2.25.1
setuptools~=56.0.0
python-dateutil~=2.8.1